fix(llm): reduce Claude no-prefill bypass discoverability, add HTTP-level test coverage
split_messages/split_messages_structured narrowed from pub(super) to
pub(in crate::claude) and the ~22 pure-conversion tests that called them
directly relocated from claude/tests.rs into request.rs's own test module,
removing the sibling-module glob-import path a future call site would
reach for. Doc comments on the splitters and on the structured_history/
plain_history funnels state explicitly that this raises bypass friction
rather than adding a compiler-enforced barrier, since Rust cannot grant a
child module's item to its parent while excluding the parent's other
children; a design that would make bypassing the funnel a compile error
is tracked separately.

Claude's Messages API base URL is now carried on an injectable api_url
field (defaulting to the existing API_URL constant) with a #[cfg(test)]
builder, replacing a dead hand-rolled TCP mock that had given up on
HTTP-level assertions. New wiremock-backed tests drive chat_with_tools,
chat_with_tools_stream, and chat_typed through a mocked HTTP round-trip
and assert the no-prefill gate strips a trailing assistant message in
the request actually sent over the wire, closing the gap where prior
coverage only exercised the identical body-construction sequence via
debug_request_json as a same-order proxy.

Closes #6155
Closes #6156
